% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{DeOPUS-package}
\alias{DeOPUS}
\alias{DeOPUS-package}
\title{DeOPUS: Deconvolution via Optimized Power-transformed Unmixing with Shrinkage}
\description{
A reference-based cellular deconvolution method that employs hierarchical
shrinkage transformation to robustly estimate cell-type proportions from
bulk RNA sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tinnlab/DeOPUS}
  \item Report bugs at \url{https://github.com/tinnlab/DeOPUS/issues}
}

}
\author{
\strong{Maintainer}: Ha Nguyen \email{hvn0006@wayne.edu}

}
